+++ src/core/OrderedList.js
@@ -1,0 +1,7 @@
+import { List } from './List.js';
+
+export class OrderedList extends List {
+    render() {
+        return `<ol>\n${this.renderItems()}\n</ol>`;
+    }
+}