+++ src/core/UnorderedList.js
@@ -1,0 +1,7 @@
+import { List } from './List.js';
+
+export class UnorderedList extends List {
+    render() {
+        return `<ul>\n${this.renderItems()}\n</ul>`;
+    }
+}